requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
slowapi
python-multipart
sqlalchemy
python-dotenv
azure-ai-inference
aiohttp
redis
docker
pytest
pylint
razorpay
numpy
duckduckgo-search
psutil